some modification

+++ Servintf.java
@@ -7,5 +7,5 @@ public interface Servintf extends Remote {
 
   void addEdge(String id, Integer u, Integer v, Integer w) throws RemoteException;
 
-  Integer getMst(String id) throws RemoteException;
+  long getMst(String id) throws RemoteException;
 }
